FIX: Move telemetry atexit into entrypoint func
The final process status is not properly sent to migas, as the exit function is not registered in the correct location.
